+++ rnalysis/gui/gui.py
@@ -1805,8 +1805,9 @@ class FilterTabPage(TabPage):
                         'split_hierarchical': 'Hierarchical (Agglomerative)', 'split_hdbscan': 'HDBSCAN',
                         'split_clicom': 'CLICOM (Ensemble)'}
     SUMMARY_FUNCS = {'describe', 'head', 'tail', 'biotypes_from_ref_table', 'biotypes_from_gtf', 'print_features'}
-    GENERAL_FUNCS = {'sort', 'transform', 'translate_gene_ids', 'differential_expression_deseq2', 'fold_change',
-                     'average_replicate_samples', 'drop_columns', 'differential_expression_limma_voom'}
+    GENERAL_FUNCS = {'sort', 'sort_by_principal_component', 'transform', 'translate_gene_ids',
+                     'differential_expression_deseq2', 'fold_change', 'average_replicate_samples', 'drop_columns',
+                     'differential_expression_limma_voom'}
     THREADED_FUNCS = {'translate_gene_ids', 'differential_expression_deseq2', 'filter_by_kegg_annotations',
                       'filter_by_go_annotations', 'differential_expression_limma_voom'}
     startedClustering = QtCore.pyqtSignal(object, object, object)